Improved Encoding & Decoding
Improved message encoding detection has been implemented which has
significntly improved the ability for mapping to the default UTF-8, so
formats such as GB2312 (Chinese) are now mapping and displaying
successfully.

In addition some work has been done to ensure that the GBP symbol (£) is
encoded correctly in UTF-8 when sent using ISO-8859-1 which was causing
issues when converted with mb_convert_encoding() and
imap_mime_header_decode().

Output encoding can now also be specified if UTF-8 is not desired.
Although, headers are still UTF-8.

Support for embedded images has also been implemeted when sent with a
content id (cid), so images can be automatically downloaded and displayed
in the email, although more needs to be done to check this works across
different mail servers.