Refactoring
The ImapMessage class has been completely refactored with the logic moved
to ImapMailManager and a Factory class has been added for creating a
message object from the response.

Collections have also been implemented for Message, Email and Folder
objects so we can check that the correct type of object array is being used.